Added warp generator, calculator, interpreter and a few placeholder screen classes
*Command interpretation has been abstracted into the static Interpreter
class, as opposed to it being used in the Ship class.
*Added the static Calculator class for working out distances and such.
*Added the WarpGenerator class, which is another Module subclass.
*Added more functions to the Module class, such as the ability to set
the operating percentage, get the operating percentage and to set the
default percentage as zero when modules are constructed. A method was
added to the Module class to get the module type name, which will later
be replaced with an integer to represent the module type.
*Added get coordinate methods to the Ship class and a temporary
getWarpGenerator method. Soon this method will change to something like
"getModule(int m)" and will return a type of module based on the
provided int.